+++ frontend/src/CardGrid/CardGrid.js
@@ -7,47 +7,30 @@ import CardContent from '@material-ui/core/CardContent';
 import CardMedia from '@material-ui/core/CardMedia';
 import Button from '@material-ui/core/Button';
 import Typography from '@material-ui/core/Typography';
+import Grid from '@material-ui/core/Grid';
+import SingleCard from './SingleCard'
 
-const styles = {
-  card: {
-    width: "40vh;"
-  },
-  media: {
-    height: 0,
-    paddingTop: '56.25%', // 16:9
-  },
-};
-
-function CardGrid(props) {
-  const { classes } = props;
-  return (
-    <div>
-      <Card className={classes.card}>
-        <CardMedia
-          className={classes.media}
-          image="/card_image_placeholder.jpeg"
-          title="Placeholder image"
-        />
-        <CardContent>
-          <Typography gutterBottom variant="headline" component="h2">
-            Hello World
-          </Typography>
-          <Typography component="p">
-            This is some secondary text.
-          </Typography>
-        </CardContent>
-        <CardActions>
-          <Button size="small" color="primary">
-            See More
-          </Button>
-        </CardActions>
-      </Card>
-    </div>
-  );
+class CardGrid extends React.Component{ //cars section (the third one)
+	render(){
+		return (
+			<Grid container spacing={24}>
+				<Grid item xs={12} sm={6}>
+					<Grid container justify="center">
+						<Grid key={1} item>
+							<SingleCard />
+						</Grid>
+					</Grid>
+				</Grid>
+				<Grid item xs={12} sm={6}>
+					<Grid container justify="center">
+						<Grid key={2} item>
+							<SingleCard />
+						</Grid>
+					</Grid>
+				</Grid>
+			</Grid>
+		);
+	}
 }
 
-CardGrid.propTypes = {
-  classes: PropTypes.object.isRequired,
-};
-
-export default withStyles(styles)(CardGrid);
+export default CardGrid;
