+++ frontend/src/SelectPage/SelectPage.js
@@ -3,7 +3,6 @@ import Select from "@material-ui/core/Select";
 import MenuItem from "@material-ui/core/MenuItem";
 import InputLabel from "@material-ui/core/InputLabel";
 import CardGrid from "../CardGrid/CardGrid";
-import Grid from '@material-ui/core/Grid';
 
 class ConfigurationSection extends React.Component{ 
 	//configutarion settings section (the second one)
@@ -21,29 +20,6 @@ class ConfigurationSection extends React.Component{
 						<MenuItem value="Type3">Type 3</MenuItem>
 				</Select>
 			</div>
-		);
-	}
-}
-
-class CardsSection extends React.Component{ //cars section (the third one)
-	render(){
-		return (
-			<Grid container spacing={24}>
-				<Grid item xs={12} sm={6}>
-					<Grid container justify="center">
-						<Grid key={1} item>
-							<CardGrid />
-						</Grid>
-					</Grid>
-				</Grid>
-				<Grid item xs={12} sm={6}>
-					<Grid container justify="center">
-						<Grid key={2} item>
-							<CardGrid />
-						</Grid>
-					</Grid>
-				</Grid>
-			</Grid>
 		);
 	}
 }
@@ -89,7 +65,7 @@ class SelectPage extends React.Component{
 					getParentConfigurationEventType={this.state.configurationEventType}/>
 				</div>
 				<div className = "CardsSection">
-					<CardsSection />
+					<CardGrid />
 				</div>
 				<div>
 					Just a test (use this props in the future): {this.props.searchString}
